Make mingw_spawn* buildable by VS 2008
The biggest chunk is the new git-compat-util.h and cache.h that are
compatible with both gcc and VS 2008.

The rest of the changes:
- sha1_file.c: use only part of the original file;
- compat/mingw.[ch]: minor changes caused by difference in header files,
  easily portable back to git (see cosmetic changes below);
- usage.c: minor changes, related to the use of NORETURN attribute, easily
  portable back to git (see cosmetic changes below);
- strbuf.c: very minor changes to fix some warnings.

There are several kinds of cosmetic changes in this commit:
- VS 2008 enforces C much stricter than gcc; for example, it does not allow
  declaring variables in the middle of a function;

- apparently, VS 2008 is much pickier about the placement of declaration
  specification (see ticktrack in compat/mingw.c);

- headers are slightly different (see struct pollfd in compat/mingw.h);

- va_copy is apparently unavailable in VS 2008;

- VS 2008 is apparently much pickier about possible loss of data and type
  mismatch, hence some additional casting [in strbuf.c];

- VS 2008 does not have __attribute__ and does not allow to use NORETURN
  in variable or parameter declarations [in usage.c].

This changes to the "borrowed" files do not break the build of git, but
no tests have been exec'd.

+++ strbuf.c
@@ -92,7 +92,7 @@ void strbuf_ltrim(struct strbuf *sb)
 
 void strbuf_tolower(struct strbuf *sb)
 {
-	int i;
+	size_t i;
 	for (i = 0; i < sb->len; i++)
 		sb->buf[i] = tolower(sb->buf[i]);
 }
@@ -204,12 +204,12 @@ void strbuf_addf(struct strbuf *sb, const char *fmt, ...)
 	va_end(ap);
 	if (len < 0)
 		die("your vsnprintf is broken");
-	if (len > strbuf_avail(sb)) {
+	if ((size_t)len > strbuf_avail(sb)) {
 		strbuf_grow(sb, len);
 		va_start(ap, fmt);
 		len = vsnprintf(sb->buf + sb->len, sb->alloc - sb->len, fmt, ap);
 		va_end(ap);
-		if (len > strbuf_avail(sb)) {
+		if ((size_t)len > strbuf_avail(sb)) {
 			die("this should not happen, your snprintf is broken");
 		}
 	}
